Refactor

+++ Keyboard/KeyboardLayout.swift
@@ -14,11 +14,11 @@ struct KeyboardLayout {
     
     let rows: [[Key]]
     
-    func key(code: Int64) -> Key? {
+    func key(code: UInt16) -> Key? {
         return keycodeToKeyDictionary[code]
     }
     
-    private var keycodeToKeyDictionary: [Int64: Key] = [
+    private var keycodeToKeyDictionary: [UInt16: Key] = [
         49: .space,
     ]
     
@@ -41,7 +41,7 @@ struct KeyboardLayout {
             [ .dollarSign,   .percentSign, .lessThanSign, .solidus,      .squareBracket, .zero,               .one,   .two,   .three, .equalsSign, ],
         ]
         
-        let keycodeRows: [[Int64]] = [
+        let keycodeRows: [[UInt16]] = [
             [ 12, 13, 14, 15, 17, 16, 32, 34, 31, 35, ],
             [  0,  1,  2,  3,  5,  4, 38, 40, 37, 41, ],
             [  6,  7,  8,  9, 11, 45, 46, 43, 47, 44, ],
